Added website module refactoring
