+++ webapp/views/views.py
@@ -1,78 +1,14 @@
 #!/usr/bin/env python3
 
-import os
-import uuid
 
-from flask import Flask
+from webapp import app, db
+from webapp.models.products import Products, ProductImages, ProductCategories
+
 from flask import request
 from flask import render_template
-from flask import redirect
 from flask import url_for
+from flask import redirect
 from flask import flash
-
-from flask_sqlalchemy import SQLAlchemy
-
-
-class Config:
-    SECRET_KEY = os.environ['BWW_SECRET_KEY']
-    DB_HOST    = os.environ.get('BWW_DB_HOST') or "localhost"
-    DB_PORT    = os.environ.get('BWW_DB_PORT') or 3306
-    DB_USER    = os.environ.get('BWW_DB_USER') or "appuser"
-    DB_PASSWD  = os.environ['BWW_DB_PASSWD']
-    DB_NAME    = os.environ.get('BWW_DB_NAME') or "BerginWoodWorking"
-    SQLALCHEMY_DATABASE_URI = f"mysql://{DB_USER}:{DB_PASSWD}@{DB_HOST}:{DB_PORT}/{DB_NAME}"
-    SQLALCHEMY_TRACK_MODIFICATIONS = False
-    DEBUG      = True
-
-
-config = Config()
-
-app = Flask(__name__)
-app.config.from_object(config)
-
-db = SQLAlchemy(app)
-
-
-class Products(db.Model):
-    __tablename__ = 'products'
-
-    product_id = db.Column(db.String(36), default=lambda: str(uuid.uuid4()), primary_key=True)
-    sku = db.Column(db.Integer)
-    product_name = db.Column(db.String(255))
-    product_category = db.Column(db.String(255))
-    product_cost = db.Column(db.Numeric(10, 2))
-    suggested_sales_price = db.Column(db.Numeric(10, 2))
-    sales_tax_payable_based_on_suggested_sales_price = db.Column(db.Numeric(10, 2))
-    profit_based_on_suggested_sales_price = db.Column(db.Numeric(10, 2))
-    percentage_of_profit = db.Column(db.String(10))
-    units_made = db.Column(db.Integer)
-    units_sold = db.Column(db.Integer)
-    units_available_for_sale = db.Column(db.Integer)
-    cost_of_units_made = db.Column(db.Numeric(10, 2))
-    cost_of_inventory_available_for_sale = db.Column(db.Numeric(10, 2))
-    cogs_per_units_sold = db.Column(db.Numeric(10, 2))
-    money_collected_including_sales_tax = db.Column(db.Numeric(10, 2))
-    date_sold = db.Column(db.Date)
-    weight = db.Column(db.Numeric(10, 2))
-    product_images = db.relationship('ProductImages', backref='product', lazy=True)
-
-
-class ProductImages(db.Model):
-    __tablename__ = 'product_images'
-    
-    product_image_id = db.Column(db.String(36), default=lambda: str(uuid.uuid4()), primary_key=True)
-    product_id = db.Column(db.String(36), db.ForeignKey('products.product_id'))
-    sku = db.Column(db.Integer)
-    encoded_image = db.Column(db.Text)
-
-
-class ProductCategories(db.Model):
-    __tablename__ = 'product_categories'
-    
-    product_category_id = db.Column(db.String(36), default=lambda: str(uuid.uuid4()), primary_key=True)
-    category_code = db.Column(db.String(10))
-    category_name = db.Column(db.String(255))
-    encoded_image = db.Column(db.Text)
 
 
 @app.route("/", methods=["GET"])
@@ -202,6 +138,3 @@ def edit_product(product_id=None):
             flash("Unable to edit product.", "danger")
 
     return render_template('edit_product.html', product=product, image=image)
-
-if __name__ == '__main__':
-    app.run(host='127.0.0.1', port=8888, debug=True)
